+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 ### Added
 - The `set_save_status()` function.
 - The `id_tag` option to `action_button_html()`.
+- Field number in error messages.
 ### Changed
 - Removed redirects after non-standard URL parameters.
+- The `add_action()` button will ask about object type if
+  `ask_object_type` is `True`.
+- The `new_window` option to `action_button_html()` can be used to set
+  the `target` of the hyperlink.
 ### Fixed
 - Field trim JavaScript was affecting file variables and triggering an
   error.